feat Buttons.js에 버튼 디자인 추가, 내 정보 수정란의 드롭다운 버그 수정
