+++ front/src/pages/MyPage/EditProfile/EditProfileMain.js
@@ -4,54 +4,10 @@ import Container from '@mui/material/Container';
 import { BasicButton } from '../../../components/Buttons';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import * as React from 'react';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemText from '@mui/material/ListItemText';
-import MenuItem from '@mui/material/MenuItem';
-import Menu from '@mui/material/Menu';
+
 import WithDrawal from './WithDrawal';
-
-const options = [
-  '연령대',
-  '10대',
-  '20대',
-  '30대',
-  '40대',
-  '50대',
-  '60대',
-  '70대',
-  '80대',
-  '90대',
-  '100대',
-  '피터팬🧚‍♀️',
-];
-
-const options2 = [
-  '선호 장르',
-  '건강/취미',
-  '경제경영',
-  '고전',
-  '과학',
-  '대학교재/전문서적',
-  '만화',
-  '사회과학',
-  '소설/시/희곡',
-  '수험서/자격증',
-  '어린이',
-  '에세이',
-  '여행',
-  '역사',
-  '요리/살림',
-  '외국어',
-  '유아',
-  '인문학',
-  '자기계발',
-  '장르소설',
-  '잡지',
-  '종교/역학',
-  '컴퓨터/모바일',
-];
+import FavoriteGenre from './FavoriteGenre';
+import AgeGroup from './AgeGroup';
 
 const TitleText = styled.div`
   width: 100%;
@@ -106,33 +62,6 @@ const Btn = styled(BasicButton)`
 `;
 
 const EditProfileMain = () => {
-  const [anchorEl, setAnchorEl] = React.useState(null);
-  const [anchorEl2, setAnchorEl2] = React.useState(null);
-  const [selectedIndex, setSelectedIndex] = React.useState(1);
-  const [selectedIndex2, setSelectedIndex2] = React.useState(1);
-  const open = Boolean(anchorEl);
-  const open2 = Boolean(anchorEl2);
-  const handleClickListItem = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClickListItem2 = (event) => {
-    setAnchorEl2(event.currentTarget);
-  };
-  const handleMenuItemClick = (event, index) => {
-    setSelectedIndex(index);
-    setAnchorEl(null);
-  };
-  const handleMenuItemClick2 = (event, index) => {
-    setSelectedIndex2(index);
-    setAnchorEl2(null);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-  const handleClose2 = () => {
-    setAnchorEl2(null);
-  };
-  const ITEM_HEIGHT = 24;
   return (
     <>
       {/* xs , sm, md, lg, xl 사이즈 */}
@@ -209,45 +138,7 @@ const EditProfileMain = () => {
           <Grid item xs={4} align="left" justifyContent="left">
             <ItemText>연령대</ItemText>
           </Grid>
-          <List
-            component="nav"
-            sx={{ bgcolor: 'background.paper', padding: 0 }}
-          >
-            <ListItem
-              button
-              id="lock-button2"
-              aria-haspopup="listbox"
-              aria-controls="lock-menu"
-              aria-expanded={open ? 'true' : undefined}
-              onClick={handleClickListItem}
-              sx={{ padding: 0 }}
-            >
-              <ListItemText secondary={options2[selectedIndex]} />
-            </ListItem>
-          </List>
-          <Menu
-            id="lock-menu"
-            anchorEl={anchorEl}
-            open={open}
-            onClose={handleClose}
-            MenuListProps={{
-              'aria-labelledby': 'lock-button2',
-              role: 'listbox',
-              maxHeight: ITEM_HEIGHT * 4.5,
-              width: '20ch',
-            }}
-          >
-            {options.map((option, index) => (
-              <MenuItem
-                key={option}
-                disabled={index === 0}
-                selected={index === selectedIndex}
-                onClick={(event) => handleMenuItemClick(event, index)}
-              >
-                {option}
-              </MenuItem>
-            ))}
-          </Menu>
+          <AgeGroup />
         </Grid>
         <ItemText>
           <Grid
@@ -262,45 +153,7 @@ const EditProfileMain = () => {
               <ItemText>선호 장르</ItemText>
             </Grid>
 
-            <List
-              component="nav"
-              sx={{ bgcolor: 'background.paper', padding: 0 }}
-            >
-              <ListItem
-                button
-                id="lock-button2"
-                aria-haspopup="listbox"
-                aria-controls="lock-menu"
-                aria-expanded={open2 ? 'true' : undefined}
-                onClick={handleClickListItem2}
-                sx={{ padding: 0 }}
-              >
-                <ListItemText secondary={options2[selectedIndex2]} />
-              </ListItem>
-            </List>
-            <Menu
-              id="lock-menu"
-              anchorEl={anchorEl2}
-              open={open2}
-              onClose={handleClose2}
-              MenuListProps={{
-                'aria-labelledby': 'lock-button2',
-                role: 'listbox',
-                maxHeight: ITEM_HEIGHT * 4.5,
-                width: '20ch',
-              }}
-            >
-              {options2.map((option2, index) => (
-                <MenuItem
-                  key={option2}
-                  disabled={index === 0}
-                  selected={index === setSelectedIndex2}
-                  onClick={(event) => handleMenuItemClick2(event, index)}
-                >
-                  {option2}
-                </MenuItem>
-              ))}
-            </Menu>
+            <FavoriteGenre />
           </Grid>
         </ItemText>
         <WithDrawal>
